SolarMeter.ino: small fix

+++ SolarMeter.ino
@@ -26,7 +26,7 @@ byte   lastMinute;
 byte   iDay;
 byte   iHour;
 byte   iMinute;
-int    upTime;               // the amount of hours the Arduino is running
+unsigned long upTime;  // the amount of hours the Arduino is running
 EthernetServer server(555);  // port changed from 80 to 555
 EthernetUDP Udp;
 char   webData[14];
